Content updated
Added content to the sections "Session Management", "Business Logic" and "Client Side". Some format needs to be checked, since tables are floating all over the place and the overflow isn't working as intended.
Added a forkBomb script, which would be executed at the end of the presentation to show a DoS attack.
Added a screenshot (still not included in the report)
Corrected a minor graphic bug found in the new_transaction.php file.